rename style sheet files
